Move AggregatedState out of 'models' package
